+++ components/Tabs/constants/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @cypress-design/constants-tabs
 
+## 0.5.1
+
+### Patch Changes
+
+- [`743a060`](https://github.com/cypress-io/cypress-design/commit/743a060b016a346a9a4ce38e630cec4b9aa24705) Thanks [@elevatebart](https://github.com/elevatebart)! - remove translate-y from center tabs
+
+## 0.5.0
+
+### Minor Changes
+
+- [#283](https://github.com/cypress-io/cypress-design/pull/283) [`cec77ef`](https://github.com/cypress-io/cypress-design/commit/cec77ef4d2e4af090efcb9e61905208b785f8183) Thanks [@elevatebart](https://github.com/elevatebart)! - add centered variant for blog
+
+### Patch Changes
+
+- Updated dependencies [[`e390827`](https://github.com/cypress-io/cypress-design/commit/e3908270ece513886c64cebcb78cebe0cd0d8c3d)]:
+  - @cypress-design/icon-registry@0.25.0
+
+## 0.4.1
+
+### Patch Changes
+
+- [`869d592`](https://github.com/cypress-io/cypress-design/commit/869d59290038d8919c318f4b6594867d1f1b8d4f) Thanks [@elevatebart](https://github.com/elevatebart)! - adust the underline-small classes with relative
+
 ## 0.4.0
 
 ### Minor Changes