+++ components/Tabs/react/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @cypress-design/react-tabs
 
+## 0.7.1
+
+### Patch Changes
+
+- [`743a060`](https://github.com/cypress-io/cypress-design/commit/743a060b016a346a9a4ce38e630cec4b9aa24705) Thanks [@elevatebart](https://github.com/elevatebart)! - remove translate-y from center tabs
+
+- Updated dependencies [[`743a060`](https://github.com/cypress-io/cypress-design/commit/743a060b016a346a9a4ce38e630cec4b9aa24705)]:
+  - @cypress-design/constants-tabs@0.5.1
+
+## 0.7.0
+
+### Minor Changes
+
+- [#283](https://github.com/cypress-io/cypress-design/pull/283) [`cec77ef`](https://github.com/cypress-io/cypress-design/commit/cec77ef4d2e4af090efcb9e61905208b785f8183) Thanks [@elevatebart](https://github.com/elevatebart)! - add centered variant for blog
+
+### Patch Changes
+
+- Updated dependencies [[`cec77ef`](https://github.com/cypress-io/cypress-design/commit/cec77ef4d2e4af090efcb9e61905208b785f8183)]:
+  - @cypress-design/constants-tabs@0.5.0
+
+## 0.6.2
+
+### Patch Changes
+
+- [`869d592`](https://github.com/cypress-io/cypress-design/commit/869d59290038d8919c318f4b6594867d1f1b8d4f) Thanks [@elevatebart](https://github.com/elevatebart)! - adust the underline-small classes with relative
+
+- Updated dependencies [[`869d592`](https://github.com/cypress-io/cypress-design/commit/869d59290038d8919c318f4b6594867d1f1b8d4f)]:
+  - @cypress-design/constants-tabs@0.4.1
+
 ## 0.6.1
 
 ### Patch Changes